[names-refactor] Delete Actions and Reducers folder from bundleLanding page.
